Add patch for the original code
Replace signal.h with musl's ucontext.h. makecontext uses the struct
ucontext_t and this struct is defined in bits/signal.h which included
from ucontext.h. In musl signal.h includes bits/signal.h but this is not
the case with newlib which has no definition for ucontext_t struct.

Signed-off-by: Charalampos Mainas <Charalampos.Mainas@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>